Version 1.0.1

+++ src/main/java/me/MathiasMC/BattleDrones/placeholders/InternalPlaceholders.java
@@ -28,30 +28,30 @@ public class InternalPlaceholders {
         return "";
     }
 
-    public long getDroneHealth(PlayerConnect playerConnect, String uuid) {
+    public int getDroneHealth(PlayerConnect playerConnect, String uuid) {
         if (playerConnect.hasActive() && plugin.listDroneHolder().contains(uuid) && plugin.getDroneHolderUUID(uuid).containsKey(playerConnect.getActive())) {
             return plugin.getDroneHolder(uuid, playerConnect.getActive()).getHealth();
         }
         return 0;
     }
 
-    public long getDroneMaxHealth(PlayerConnect playerConnect, String uuid) {
+    public int getDroneMaxHealth(PlayerConnect playerConnect, String uuid) {
         if (playerConnect.hasActive()  && plugin.listDroneHolder().contains(uuid) && plugin.getDroneHolderUUID(uuid).containsKey(playerConnect.getActive())) {
             DroneHolder droneHolder = plugin.getDroneHolder(uuid, playerConnect.getActive());
-            return plugin.droneFiles.get(playerConnect.getActive()).getLong(playerConnect.getGroup() + "." + droneHolder.getLeft() + ".health");
+            return plugin.droneFiles.get(playerConnect.getActive()).getInt(playerConnect.getGroup() + "." + droneHolder.getLeft() + ".health");
         }
         return 0;
     }
 
-    public long getDroneMaxAmmo(PlayerConnect playerConnect, String uuid) {
+    public int getDroneMaxAmmo(PlayerConnect playerConnect, String uuid) {
         if (playerConnect.hasActive()  && plugin.listDroneHolder().contains(uuid) && plugin.getDroneHolderUUID(uuid).containsKey(playerConnect.getActive())) {
             DroneHolder droneHolder = plugin.getDroneHolder(uuid, playerConnect.getActive());
-            return plugin.droneFiles.get(playerConnect.getActive()).getLong(playerConnect.getGroup() + "." + droneHolder.getLevel() + ".max-ammo-slots") * 64;
+            return plugin.droneFiles.get(playerConnect.getActive()).getInt(playerConnect.getGroup() + "." + droneHolder.getLevel() + ".max-ammo-slots") * 64;
         }
         return 0;
     }
 
-    public long getDroneAmmo(PlayerConnect playerConnect, String uuid) {
+    public int getDroneAmmo(PlayerConnect playerConnect, String uuid) {
         if (playerConnect.hasActive()  && plugin.listDroneHolder().contains(uuid) && plugin.getDroneHolderUUID(uuid).containsKey(playerConnect.getActive())) {
             DroneHolder droneHolder = plugin.getDroneHolder(uuid, playerConnect.getActive());
             return droneHolder.getAmmo();
